mimeo #33 - refactor mimeo renderers module file name
