+++ src/mimeo/utils/__init__.py
@@ -2,4 +2,4 @@ from .mimeo_utils import (AutoIncrementUtil, CityUtil, CountryUtil,
                           CurrentIterationUtil, DateTimeUtil, DateUtil,
                           FirstNameUtil, KeyUtil, LastNameUtil, MimeoUtil,
                           RandomIntegerUtil, RandomItemUtil, RandomStringUtil)
-from .renderer import MimeoRenderer
+from .renderers import MimeoRenderer
