+++ tests/utils/test_date_util.py
@@ -1,6 +1,6 @@
 from datetime import date, timedelta
 
-from mimeo.utils.renderer import UtilsRenderer
+from mimeo.utils.renderers import UtilsRenderer
 
 
 def test_date_raw():
