+++ tests/utils/test_special_fields_renderer.py
@@ -3,7 +3,7 @@ import pytest
 from mimeo.config import MimeoConfig
 from mimeo.context import MimeoContextManager
 from mimeo.utils.exc import NotASpecialField
-from mimeo.utils.renderer import SpecialFieldsRenderer
+from mimeo.utils.renderers import SpecialFieldsRenderer
 
 
 @pytest.fixture(autouse=True)
